[Customer Crud] Implement completely

+++ src/controller/user.controller.ts
@@ -1,4 +1,6 @@
 import { NextFunction, Request, Response } from 'express';
+import { __ } from 'i18n';
+import { User } from '../entity/user';
 import { ICrudController } from '../lib/ICrudController';
 import { UserService } from '../service/user.service';
 
@@ -18,6 +20,10 @@ class UserController implements ICrudController {
     }
 
     public create(req: Request, res: Response, next: NextFunction): void {
+        if (!UserService.checkRoleLevel((req['user'] as User).id, req.body.roles)) {
+            return next(new Error(__('user.can_not_create_user_with_higher_level')));
+        }
+
         UserService.create(req.body.username, req.body.email, req.body.password, req.body.phoneNumber, req.body.address,
             req.body.fullName, req.body.avatar, req.body.birthday, req.body.roles).then(value => {
                 return res.status(200).json(value);
@@ -31,6 +37,10 @@ class UserController implements ICrudController {
     }
 
     public update(req: Request, res: Response, next: NextFunction): void {
+        if (!UserService.checkRoleLevel((req['user'] as User).id, req.body.roles)) {
+            return next(new Error(__('user.can_not_update_user_with_higher_level')));
+        }
+
         UserService.edit(parseInt(req.params.id, 10), req.body.password, req.body.phoneNumber, req.body.address,
             req.body.fullName, req.body.avatar, req.body.birthday, req.body.roles).then(value =>
                 res.status(200).json(value)
@@ -38,6 +48,10 @@ class UserController implements ICrudController {
     }
 
     public delete(req: Request, res: Response, next: NextFunction): void {
+        if (!UserService.checkRoleLevel((req['user'] as User).id, req.body.roles)) {
+            return next(new Error(__('user.can_not_delete_user_with_higher_level')));
+        }
+
         UserService.delete(parseInt(req.params.id, 10)).then(() =>
             res.sendStatus(200)
         ).catch(e => next(e));
@@ -47,3 +61,4 @@ class UserController implements ICrudController {
 const userController = new UserController();
 
 export { userController as UserController };
+
